Fix edge case when comparing min_cost and leaf_cost
Sometimes, it can happen that the min_cost computed actually corresponds to the same cost than building a leaf with all the given lights. However, due to some numerical computation, the compute min_cost is not exactly equal to the leaf_cost as it should be. In that case, if we enter the condition min_cost < leaf_cost, we can have a infinite loop leading to stack overflow because we try to split an infinite number of times the primitives although they are never really split in the end. This can happen only when the minimum cost is identical along every dimensions for any split because any dimension can be chosen meaning that a dimension where all lights are at the same coordinates can be picked, leading to choosing a mid point is either equal to start or end.